Change approach to loading the default IPromptTemplateFactory
