+++ dotnet/src/SemanticKernel.Core/Extensions/KernelSemanticFunctionExtensions.cs
@@ -4,8 +4,6 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.IO;
-using System.Linq;
-using System.Reflection;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Microsoft.SemanticKernel.AI;
@@ -126,7 +124,7 @@ public static class KernelSemanticFunctionExtensions
         Verify.ValidFunctionName(functionName);
         if (!string.IsNullOrEmpty(pluginName)) { Verify.ValidPluginName(pluginName); }
 
-        var factory = promptTemplateFactory ?? CreateDefaultPromptTemplateFactory(kernel);
+        var factory = promptTemplateFactory ?? AggregatorPromptTemplateFactory.CreateDefaultPromptTemplateFactory(kernel);
         IPromptTemplate promptTemplateInstance = factory.Create(promptTemplate, promptTemplateConfig);
 
         // TODO: manage overwrites, potentially error out
@@ -246,7 +244,7 @@ public static class KernelSemanticFunctionExtensions
         const string ConfigFile = "config.json";
         const string PromptFile = "skprompt.txt";
 
-        var factory = promptTemplateFactory ?? CreateDefaultPromptTemplateFactory(kernel);
+        var factory = promptTemplateFactory ?? AggregatorPromptTemplateFactory.CreateDefaultPromptTemplateFactory(kernel);
         var functions = new Dictionary<string, ISKFunction>();
 
         ILogger? logger = null;
@@ -310,92 +308,6 @@ public static class KernelSemanticFunctionExtensions
             promptTemplate,
             kernel.LoggerFactory
         );
-    }
-
-    private const string BasicTemplateFactoryAssemblyName = "Microsoft.SemanticKernel.TemplateEngine.Basic";
-    private const string BasicTemplateFactoryTypeName = "BasicPromptTemplateFactory";
-    private static bool s_promptTemplateFactoryInitialized = false;
-    private static Type? s_promptTemplateFactoryType = null;
-
-    /// <summary>
-    /// Create a default prompt template factory.
-    ///
-    /// This is a temporary solution to avoid breaking existing clients.
-    /// There will be a separate task to add support for registering instances of IPromptTemplateEngine and obsoleting the current approach.
-    ///
-    /// </summary>
-    /// <param name="kernel">Kernel instance</param>
-    /// <returns>Instance of <see cref="IPromptTemplateEngine"/>.</returns>
-    private static IPromptTemplateFactory CreateDefaultPromptTemplateFactory(IKernel kernel)
-    {
-#pragma warning disable CS0618 // Type or member is obsolete
-        if (kernel.PromptTemplateEngine is not null)
-        {
-            return new PromptTemplateFactory(kernel.PromptTemplateEngine);
-        }
-#pragma warning restore CS0618 // Type or member is obsolete
-
-        if (!s_promptTemplateFactoryInitialized)
-        {
-            s_promptTemplateFactoryType = GetPromptTemplateFactoryType();
-            s_promptTemplateFactoryInitialized = true;
-        }
-
-        if (s_promptTemplateFactoryType is not null)
-        {
-            var constructor = s_promptTemplateFactoryType.GetConstructor(new Type[] { typeof(ILoggerFactory) });
-            if (constructor is not null)
-            {
-#pragma warning disable CS8601 // Null logger factory is OK
-                var factory = (IPromptTemplateFactory)constructor.Invoke(new object[] { kernel.LoggerFactory });
-                if (factory is not null)
-                {
-                    return factory;
-                }
-#pragma warning restore CS8601
-            }
-        }
-
-        throw new SKException($"Unable to create default prompt template factory. Please provide an implementation of IPromptTemplateFactory or depend on {BasicTemplateFactoryAssemblyName}");
-    }
-
-    /// <summary>
-    /// Get the prompt template engine type if available
-    /// </summary>
-    /// <returns>The type for the prompt template engine if available</returns>
-    private static Type? GetPromptTemplateFactoryType()
-    {
-        try
-        {
-            var assembly = Assembly.Load(BasicTemplateFactoryAssemblyName);
-
-            return assembly.ExportedTypes.Single(type =>
-                type.Name.Equals(BasicTemplateFactoryTypeName, StringComparison.Ordinal) &&
-                type.GetInterface(nameof(IPromptTemplateFactory)) is not null);
-        }
-        catch (Exception ex) when (!ex.IsCriticalException())
-        {
-            return null;
-        }
-    }
-
-    #endregion
-
-    #region Obsolete
-    [Obsolete("IPromptTemplateEngine is being replaced with IPromptTemplateFactory. This will be removed in a future release.")]
-    internal sealed class PromptTemplateFactory : IPromptTemplateFactory
-    {
-        private readonly IPromptTemplateEngine _promptTemplateEngine;
-
-        public PromptTemplateFactory(IPromptTemplateEngine promptTemplateEngine)
-        {
-            this._promptTemplateEngine = promptTemplateEngine;
-        }
-
-        public IPromptTemplate Create(string templateString, PromptTemplateConfig promptTemplateConfig)
-        {
-            return new PromptTemplate(templateString, promptTemplateConfig, this._promptTemplateEngine);
-        }
     }
     #endregion
 }
